+++ library/aik099/PHPUnit/BrowserConfiguration/BrowserStackBrowserConfiguration.php
@@ -14,7 +14,6 @@ namespace aik099\PHPUnit\BrowserConfiguration;
 use aik099\PHPUnit\APIClient\BrowserStackAPIClient;
 use aik099\PHPUnit\APIClient\IAPIClient;
 use aik099\PHPUnit\BrowserTestCase;
-use aik099\PHPUnit\Event\TestEvent;
 use WebDriver\ServiceFactory;
 
 /**
@@ -41,19 +40,11 @@ class BrowserStackBrowserConfiguration extends ApiBrowserConfiguration
 	}
 
 	/**
-	 * Hook, called from "BrowserTestCase::setUp" method.
-	 *
-	 * @param TestEvent $event Test event.
-	 *
-	 * @return void
+	 * @inheritDoc
 	 */
-	public function onTestSetup(TestEvent $event)
+	public function onTestSetup(BrowserTestCase $test_case)
 	{
-		if ( !$event->validateSubscriber($this->getTestCase()) ) {
-			return;
-		}
-
-		parent::onTestSetup($event);
+		parent::onTestSetup($test_case);
 
 		$desired_capabilities = $this->getDesiredCapabilities();